Add test related to registerRenderer
An error should be thrown if serverRenderReactComponent tries to
render a component that was registered with registerRenderer. This
is because registerRenderer is for use cases where you need to
wait before rendering, and this is not possible in the ExecJS
environment.

Any components on the server side should be registered with the
normal register API, probably by having seperate bundles for
server and client with different entry points.