Add an integration test for registerRenderer
Added a new page /client_side_manual_render. This page renders the
react component `ManualRenderApp`, which is registered on the client
side using `registerRenderer`. The given renderer function just
renders some simple html to the DOM.

The integration test tests for the presence of the html rendered by
the ManualRenderApp renderer. The test passing suggests that
ReactOnRails properly delegates to the renderer function on client
startup.